requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
blinker=1.4=py_1
ca-certificates=2019.11.28=hecc5488_0
cachetools=3.1.1=py_0
certifi=2019.11.28=py37hc8dfbb8_1
cffi=1.14.0=py37ha419a9e_0
chardet=3.0.4=py37hc8dfbb8_1006
click=7.1.1=pyh8c360ce_0
cryptography=2.8=py37hb32ad35_1
cycler=0.10.0=py_2
freetype=2.10.1=ha9979f8_0
google-api-core=1.16.0=py37h21ff451_1
google-api-python-client=1.8.0=pyh8c360ce_0
google-auth=1.11.2=py_0
google-auth-httplib2=0.0.3=py_3
google-auth-oauthlib=0.4.1=py_2
googleapis-common-protos=1.51.0=py37h21ff451_1
httplib2=0.17.0=py37hc8dfbb8_1
icc_rt=2019.0.0=h0cc432a_1
icu=64.2=he025d50_1
idna=2.9=py_1
intel-openmp=2020.0=166
jpeg=9c=hfa6e2cd_1001
kiwisolver=1.1.0=py37heaa310e_1
libblas=3.8.0=15_mkl
libcblas=3.8.0=15_mkl
libclang=9.0.1=default_hf44288c_0
liblapack=3.8.0=15_mkl
libpng=1.6.37=hfe6a214_1
libprotobuf=3.11.4=h1a1b453_0
libtiff=4.1.0=h885aae3_6
lz4-c=1.8.3=he025d50_1001
matplotlib=3.2.1=0
matplotlib-base=3.2.1=py37h911224e_0
mkl=2020.0=166
numpy=1.18.1=py37hc71023c_0
oauthlib=3.0.1=py_0
olefile=0.46=py_0
openssl=1.1.1e=hfa6e2cd_0
pillow=7.0.0=py37h91e7a8d_1
pip=20.0.2=py_2
protobuf=3.11.4=py37he025d50_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=py_0
pyjwt=1.7.1=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.6=py_0
pyqt=5.12.3=py37h6538335_1
pyqt5-sip=4.19.18=pypi_0
pyqtwebengine=5.12.1=pypi_0
pysocks=1.7.1=py37hc8dfbb8_1
python=3.7.6=h5b45d93_4_cpython
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytz=2019.3=py_0
qt=5.12.5=h7ef1ec2_0
requests=2.23.0=pyh8c360ce_2
requests-oauthlib=1.2.0=py_0
rsa=4.0=py_0
scipy=1.1.0=py37h29ff71c_2
setuptools=46.1.1=py37hc8dfbb8_0
simplejson=3.17.0=py37hfa6e2cd_0
six=1.14.0=py_1
sqlite=3.30.1=hfa6e2cd_0
tk=8.6.10=hfa6e2cd_0
tornado=6.0.4=py37hfa6e2cd_0
uritemplate=3.0.1=py_0
urllib3=1.25.7=py37hc8dfbb8_1
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_1
wheel=0.34.2=py_1
win_inet_pton=1.1.0=py37_0
wincertstore=0.2=py37_1003
xz=5.2.4=h2fa13f4_1002
zlib=1.2.11=h2fa13f4_1006
zstd=1.4.4=hd8a0e53_2
